Adding a _common_replacements dictionary. For a "to Qt.py converter".
So that a "convert from X to Qt.py tool" can use them similar to
_common_members.

Currently it is not storing the lambda functions in the dict.
I am thinking on support for it currently.

Three new private functions have been added to automate the
functionality that was there before.
These functions are _setattr, _getattr and _set_commond_replacements.
Their docstrings should hopefully aid in what their purposes are.

## *To note*
_setattr is currently going to add a dummy ModuleType if part of the
path is missing. I want to discuss what the desired functionality may be
 for this. The "creating ModuleTypes" is not very "explicit".

_getattr has a similar thing, if an attribute from the source side is
not available, or missing, it raises a ModuleMissingException.
_set_common_replacements catches this and throws it away. I'd imagine
there is some sort of message/warning we should give the user. Let them
know that their <binding> installation is missing modules.

I have mapped the assignments from each _<binding> function into the
_common_replacements dictionary except for the lambdas stated above.